Add a proof of concept icon reader project in c++
The plan was to make this in Python but winapi stuff is too opaque to get that working. Doing this in C++ turned out to be hard enough as it is.

The current implementation dumps all icons to bmp files found the dll or exe specified by the first command line argument